Move low-level plex functionality to its own package. Make some functions private. Few naming conventions workaround.
