+++ resources/lib/plexserver/__init__.py
@@ -1,0 +1,4 @@
+__author__ = 'oburdun'
+
+from plex_server import PlexMediaServer
+from plex_gdm import PlexGdm
